tctl resource: convert saml_idp_service_provider handler
