+++ tool/tctl/common/resource_command.go
@@ -33,7 +33,6 @@ import (
 	"time"
 
 	"github.com/alecthomas/kingpin/v2"
-	"github.com/crewjam/saml/samlsp"
 	"github.com/gravitational/trace"
 	kyaml "k8s.io/apimachinery/pkg/util/yaml"
 
@@ -131,7 +130,6 @@ func (rc *ResourceCommand) Initialize(app *kingpin.Application, _ *tctlcfg.Globa
 		types.KindNetworkRestrictions:         rc.createNetworkRestrictions,
 		types.KindKubernetesCluster:           rc.createKubeCluster,
 		types.KindLoginRule:                   rc.createLoginRule,
-		types.KindSAMLIdPServiceProvider:      rc.createSAMLIdPServiceProvider,
 		types.KindDevice:                      rc.createDevice,
 		types.KindOktaImportRule:              rc.createOktaImportRule,
 		types.KindIntegration:                 rc.createIntegration,
@@ -776,46 +774,6 @@ func (rc *ResourceCommand) createLoginRule(ctx context.Context, client *authclie
 	return nil
 }
 
-func (rc *ResourceCommand) createSAMLIdPServiceProvider(ctx context.Context, client *authclient.Client, raw services.UnknownResource) error {
-	// Create services.SAMLIdPServiceProvider from raw YAML to extract the service provider name.
-	sp, err := services.UnmarshalSAMLIdPServiceProvider(raw.Raw, services.DisallowUnknown())
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	if sp.GetEntityDescriptor() != "" {
-		// verify that entity descriptor parses
-		ed, err := samlsp.ParseMetadata([]byte(sp.GetEntityDescriptor()))
-		if err != nil {
-			return trace.BadParameter("invalid entity descriptor for SAML IdP Service Provider %q: %v", sp.GetEntityID(), err)
-		}
-
-		// issue warning about unsupported ACS bindings.
-		if err := services.FilterSAMLEntityDescriptor(ed, false /* quiet */); err != nil {
-			slog.WarnContext(ctx, "Entity descriptor for SAML IdP service provider contains unsupported ACS bindings",
-				"entity_id", sp.GetEntityID(),
-				"error", err,
-			)
-		}
-	}
-
-	serviceProviderName := sp.GetName()
-
-	exists := false
-	if err = client.CreateSAMLIdPServiceProvider(ctx, sp); err != nil {
-		if trace.IsAlreadyExists(err) {
-			exists = true
-			err = client.UpdateSAMLIdPServiceProvider(ctx, sp)
-		}
-
-		if err != nil {
-			return trace.Wrap(err)
-		}
-	}
-	fmt.Printf("SAML IdP service provider %q has been %s\n", serviceProviderName, UpsertVerb(exists, rc.IsForced()))
-	return nil
-}
-
 func (rc *ResourceCommand) createDevice(ctx context.Context, client *authclient.Client, raw services.UnknownResource) error {
 	res, err := services.UnmarshalDevice(raw.Raw)
 	if err != nil {
@@ -1204,11 +1162,6 @@ func (rc *ResourceCommand) Delete(ctx context.Context, client *authclient.Client
 			return trail.FromGRPC(err)
 		}
 		fmt.Printf("login rule %q has been deleted\n", rc.ref.Name)
-	case types.KindSAMLIdPServiceProvider:
-		if err := client.DeleteSAMLIdPServiceProvider(ctx, rc.ref.Name); err != nil {
-			return trace.Wrap(err)
-		}
-		fmt.Printf("SAML IdP service provider %q has been deleted\n", rc.ref.Name)
 	case types.KindDevice:
 		remote := client.DevicesClient()
 		device, err := findDeviceByIDOrTag(ctx, remote, rc.ref.Name)
@@ -1667,20 +1620,6 @@ func (rc *ResourceCommand) getCollection(ctx context.Context, client *authclient
 			Name: rc.ref.Name,
 		})
 		return &loginRuleCollection{[]*loginrulepb.LoginRule{rule}}, trail.FromGRPC(err)
-	case types.KindSAMLIdPServiceProvider:
-		if rc.ref.Name != "" {
-			serviceProvider, err := client.GetSAMLIdPServiceProvider(ctx, rc.ref.Name)
-			if err != nil {
-				return nil, trace.Wrap(err)
-			}
-			return &samlIdPServiceProviderCollection{serviceProviders: []types.SAMLIdPServiceProvider{serviceProvider}}, nil
-		}
-
-		resources, err := stream.Collect(clientutils.Resources(ctx, client.ListSAMLIdPServiceProviders))
-		if err != nil {
-			return nil, trace.Wrap(err)
-		}
-		return &samlIdPServiceProviderCollection{serviceProviders: resources}, nil
 	case types.KindDevice:
 		remote := client.DevicesClient()
 		if rc.ref.Name != "" {
